requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pyyaml
playwright